refactor: move Subsonic auth flow into AuthenticationService
